sec-research/requirements.txt
# Dependencies for sec-research/ workspace.
# Install: pip install -r requirements.txt   (or)   uv pip install -r requirements.txt

jsonschema>=4.20
pyyaml>=6.0
keyring>=24.0

# Test dependencies (install with --extra test if you split deps later)
pytest>=8.0
